chore: rename module to rpo_falcon_512_multisig
